Changes before error encountered
Agent-Logs-Url: https://github.com/SimulatorLife/GMLoop/sessions/6521c93c-9a04-46ed-ab89-2bc3b00b281a
