+++ src/lint/src/rules/gml/index.ts
@@ -1,2 +1,2 @@
 export { createGmlRule } from "./create-gml-rules.js";
-export type { GmlRuleDefinition } from "./rule-definition.js";
+export type { GmlRuleDefinition } from "./rule-base-helpers.js";
