+++ src/lint/src/rules/gml/rules/no-empty-regions-rule.ts
@@ -5,9 +5,9 @@ import {
     applySourceTextEdits,
     createMeta,
     reportProgramTextRewrite,
-    type SourceTextEdit
+    type SourceTextEdit,
+    type GmlRuleDefinition
 } from "../rule-base-helpers.js";
-import type { GmlRuleDefinition } from "../rule-definition.js";
 
 type RegionBlockRecord = Readonly<{
     startLineIndex: number;
