+++ src/lint/src/rules/gml/rules/no-unary-plus-on-identifier-rule.ts
@@ -1,8 +1,7 @@
 import { Core } from "@gmloop/core";
 import type { Rule } from "eslint";
 
-import { createMeta, isAstNodeRecord } from "../rule-base-helpers.js";
-import type { GmlRuleDefinition } from "../rule-definition.js";
+import { createMeta, isAstNodeRecord, type GmlRuleDefinition } from "../rule-base-helpers.js";
 
 /**
  * Unwraps chains of `ParenthesizedExpression` nodes to retrieve the innermost
