+++ src/lint/src/rules/gml/rules/optimize-logical-flow-rule.ts
@@ -2,8 +2,7 @@ import { Core, type MutableGameMakerAstNode } from "@gmloop/core";
 import type { Rule } from "eslint";
 
 import { printNodeForAutofix } from "../../../language/autofix-printing.js";
-import { createMeta, findPreviousNonWhitespaceIndex, resolveLocFromIndex } from "../rule-base-helpers.js";
-import type { GmlRuleDefinition } from "../rule-definition.js";
+import { createMeta, findPreviousNonWhitespaceIndex, resolveLocFromIndex, type GmlRuleDefinition } from "../rule-base-helpers.js";
 import { applyLogicalNormalizationWithChangeMetadata } from "../transforms/logical-expression-traversal-normalization.js";
 
 /**
