+++ src/lint/src/rules/gml/rules/prefer-increment-decrement-operators-rule.ts
@@ -7,9 +7,9 @@ import {
     isAstNodeRecord,
     isStandaloneStatementParentKey,
     sourceRangeContainsCommentToken,
-    walkAstNodesWithParent
+    walkAstNodesWithParent,
+    type GmlRuleDefinition
 } from "../rule-base-helpers.js";
-import type { GmlRuleDefinition } from "../rule-definition.js";
 
 type IncrementDecrementAssignmentOperator = "+=" | "-=";
 type IncrementDecrementOperator = "++" | "--";
