+++ src/lint/src/rules/gml/rules/remove-default-comments-rule.ts
@@ -1,8 +1,7 @@
 import { Core } from "@gmloop/core";
 import type { Rule } from "eslint";
 
-import { createMeta, reportProgramTextRewrite } from "../rule-base-helpers.js";
-import type { GmlRuleDefinition } from "../rule-definition.js";
+import { createMeta, reportProgramTextRewrite, type GmlRuleDefinition } from "../rule-base-helpers.js";
 
 const DEFAULT_COMMENT_PLACEHOLDER_FRAGMENTS = Object.freeze([
     "Script assets have changed for v2.3.0",
