+++ src/lint/src/rules/gml/rules/require-trailing-optional-defaults-rule.ts
@@ -9,9 +9,9 @@ import {
     isAstNodeRecord,
     reportFullTextRewrite,
     type SourceTextEdit,
-    walkAstNodes
+    walkAstNodes,
+    type GmlRuleDefinition
 } from "../rule-base-helpers.js";
-import type { GmlRuleDefinition } from "../rule-definition.js";
 
 const { getNodeStartIndex, getNodeEndIndex, unwrapParenthesizedExpression: unwrapParenthesized } = Core;
