+++ src/lint/src/rules/gml/rules/simplify-real-calls-rule.ts
@@ -29,8 +29,7 @@
 import { Core } from "@gmloop/core";
 import type { Rule } from "eslint";
 
-import { createMeta, isAstNodeRecord, walkAstNodes } from "../rule-base-helpers.js";
-import type { GmlRuleDefinition } from "../rule-definition.js";
+import { createMeta, isAstNodeRecord, walkAstNodes, type GmlRuleDefinition } from "../rule-base-helpers.js";
 
 /**
  * Matches a valid GML numeric literal string: optional sign, integer part,
